Handle creator and contributor multi-part fields for collections
