+++ .rubocop.yml
@@ -34,6 +34,7 @@ Metrics/BlockLength:
     - 'app/models/concerns/hyku_addons/generic_work_overrides.rb'
     - 'app/models/concerns/hyku_addons/solr_document_behavior.rb'
     - 'app/models/concerns/hyku_addons/work_base.rb'
+    - 'app/forms/concerns/hyku_addons/collection_form_behavior.rb'
     - 'app/forms/concerns/hyku_addons/work_form.rb'
     - 'app/models/pacific_article.rb'
     - 'app/services/bolognese/writers/ris_writer_behavior.rb'
