+++ app/forms/concerns/hyku_addons/work_form.rb
@@ -3,14 +3,7 @@ module HykuAddons
   module WorkForm
     extend ActiveSupport::Concern
 
-    # Helper methods for JSON fields
-    def creator_list
-      person_or_organization_list(:creator)
-    end
-
-    def contributor_list
-      person_or_organization_list(:contributor)
-    end
+    include HykuAddons::PersonOrOrganizationFormBehavior
 
     def editor_list
       person_or_organization_list(:editor)
@@ -21,8 +14,6 @@ module HykuAddons
       def build_permitted_params
         super.tap do |permitted_params|
           permitted_params << common_fields
-          permitted_params << creator_fields
-          permitted_params << contributor_fields
           permitted_params << date_published_fields
           permitted_params << date_accepted_fields
           permitted_params << date_submitted_fields
@@ -69,23 +60,6 @@ module HykuAddons
            library_of_congress_classification add_info issn isbn eissn version_number series_name book_title pagination
            publisher place_of_publication journal_title alternative_journal_title volume edition issue article_num
            qualification_name qualification_level]
-      end
-
-      def creator_fields
-        {
-          creator: [:creator_organization_name, :creator_given_name, :creator_middle_name, :creator_family_name, :creator_name_type,
-                    :creator_orcid, :creator_isni, :creator_ror, :creator_grid, :creator_wikidata, :creator_suffix, :creator_institution,
-                    creator_institutional_relationship: []]
-        }
-      end
-
-      def contributor_fields
-        {
-          contributor: [:contributor_organization_name, :contributor_given_name, :contributor_middle_name, :contributor_family_name,
-                        :contributor_name_type, :contributor_orcid, :contributor_isni, :contributor_ror,
-                        :contributor_grid, :contributor_wikidata, :creator_suffix, :creator_institution, :contributor_type,
-                        contributor_institutional_relationship: []]
-        }
       end
 
       def date_accepted_fields
@@ -139,13 +113,5 @@ module HykuAddons
         ]
       end
     end
-
-    private
-
-      def person_or_organization_list(field)
-        # Return empty hash to ensure that it gets rendered at least once
-        return [{}] unless send(field)&.first.present?
-        JSON.parse(send(field).first)
-      end
   end
 end
