+++ app/views/records/show_fields/_creator.html.erb
@@ -1,0 +1,1 @@
+<%= @presenter.attribute_to_html(:creator, render_as: :linked_name, values: @presenter.creator_list, search_field: 'creator_display_ssim', html_dl: true) %>
